+++ shanoir-ng-front/src/app/app.module.ts
@@ -259,6 +259,7 @@ import { LoginGuard } from "./shared/roles/login-guard";
 import { AccessRequestService } from './users/access-request/access-request.service';
 import { AccessRequestListComponent } from './users/access-request/access-request-list.component';
 import { MassDownloadService } from './shared/mass-download/mass-download.service';
+import { SingleDownloadService } from './shared/mass-download/single-download.service';
 import { ClipboardModule } from '@angular/cdk/clipboard';
 import { TaskStatusComponent } from './async-tasks/status/task-status.component';
 import { DatasetCopyDialogComponent } from "./shared/components/dataset-copy-dialog/dataset-copy-dialog.component";
@@ -548,6 +549,7 @@ import { DoubleAwesomeComponent } from './shared/double-awesome/double-awesome.c
         QualityCardService,
         QualityCardDTOService,
         MassDownloadService,
+        SingleDownloadService,
         SessionService,
         ShanoirEventService,
         TreeService,
