+++ shanoir-ng-front/src/app/examinations/shared/examination.service.ts
@@ -11,11 +11,12 @@
  * You should have received a copy of the GNU General Public License
  * along with this program. If not, see https://www.gnu.org/licenses/gpl-3.0.html
  */
-import { HttpClient, HttpResponse } from '@angular/common/http';
-import { Injectable, OnDestroy } from '@angular/core';
+import { HttpClient } from '@angular/common/http';
+import { Injectable } from '@angular/core';
 import { Observable } from 'rxjs';
 
-import { TaskState, TaskStatus } from 'src/app/async-tasks/task.model';
+import { TaskState } from 'src/app/async-tasks/task.model';
+import { SingleDownloadService } from 'src/app/shared/mass-download/single-download.service';
 import { EntityService } from '../../shared/components/entity/entity.abstract.service';
 import { Page, Pageable } from '../../shared/components/table/pageable.model';
 import * as AppUtils from '../../utils/app.utils';
@@ -23,7 +24,6 @@ import { ServiceLocator } from '../../utils/locator.service';
 import { ExaminationDTO, ExaminationDTOService } from './examination.dto';
 import { Examination } from './examination.model';
 import { SubjectExamination } from './subject-examination.model';
-import { MassDownloadService } from 'src/app/shared/mass-download/mass-download.service';
 
 
 @Injectable()
@@ -31,7 +31,7 @@ export class ExaminationService extends EntityService<Examination> {
 
     API_URL = AppUtils.BACKEND_API_EXAMINATION_URL;
 
-    constructor(protected http: HttpClient, private downloadService: MassDownloadService) {
+    constructor(protected http: HttpClient, private downloadService: SingleDownloadService) {
         super(http)
     }
     protected examinationDtoService: ExaminationDTOService = ServiceLocator.injector.get(ExaminationDTOService);
