+++ shanoir-ng-front/src/app/shared/mass-download/mass-download.service.ts
@@ -13,26 +13,25 @@
  */
 
 import { formatDate } from '@angular/common';
-import { HttpParams, HttpResponse } from '@angular/common/http';
+import { HttpResponse } from '@angular/common/http';
 import { ComponentRef, Injectable } from '@angular/core';
+import { AngularDeviceInformationService } from 'angular-device-information';
+import { Observable, race, Subscription } from 'rxjs';
 import { last, map, take } from 'rxjs/operators';
 import { Task, TaskState } from 'src/app/async-tasks/task.model';
 import { Dataset } from 'src/app/datasets/shared/dataset.model';
 import { DatasetService, Format } from 'src/app/datasets/shared/dataset.service';
+import { getSizeStr, StrictUnion } from 'src/app/utils/app.utils';
 import { ServiceLocator } from 'src/app/utils/locator.service';
 import { SuperPromise } from 'src/app/utils/super-promise';
 import { ConfirmDialogService } from '../components/confirm-dialog/confirm-dialog.service';
 import { ConsoleService } from '../console/console.service';
+import { ShanoirError } from '../models/error.model';
 import { NotificationsService } from '../notifications/notifications.service';
+import { SessionService } from '../services/session.service';
 import { DownloadSetupAltComponent } from './download-setup-alt/download-setup-alt.component';
 import { DownloadSetupComponent } from './download-setup/download-setup.component';
 import { Queue } from './queue.model';
-import { SessionService } from '../services/session.service';
-import { ShanoirError } from '../models/error.model';
-import { StrictUnion, getSizeStr } from 'src/app/utils/app.utils';
-import { AngularDeviceInformationService } from 'angular-device-information';
-import { Observable, race, Subscription } from 'rxjs';
-import * as AppUtils from '../../utils/app.utils';
 
 declare var JSZip: any;
 
@@ -620,50 +619,7 @@ export class MassDownloadService {
             this.consoleService.log('error', 'Can\'t parse the status from the recorded message', [e, task?.report]);
             return null;
         }
-    }
-
-    /**
-     * Handles large files. Download a single file, displaying a loading bar in the side menu if the dl takes more than 5s.
-     * 
-     * It is impossible to nicely use the browser dl for large files.
-     * It's because the browser need a dl through a <a href> to do so, and we can't as we have an auth token to put in a http header.
-     * But using js makes the browser dl the file silently, then copying the blob to the dl dir with the nice browser display.
-     * So with large files, the silent step is confusing for users. Here we will help her/him by displaying the progress.
-     * 
-     * @param url 
-     * @param params 
-     * @param state 
-     * @param totalSize total size of the file, if known
-     * @returns 
-     */
-    downloadSingleFile(url: string, params?: HttpParams, state?: TaskState): Observable<TaskState> {
-        let obs: Observable<TaskState> = AppUtils.downloadWithStatusGET(url, params, state);
-
-        let task: Task = new Task();
-        task.id = Date.now();
-        task.creationDate = new Date();
-        task.lastUpdate = task.creationDate;
-        task.message = 'Downloading ' + (url + '').split('/').pop();
-        task.progress = 0;
-        task.status = 2;
-        task.eventType = 'downloadFile.event';
-        task.sessionId = this.sessionService.sessionId;
-
-        let startTs: number = Date.now();
-        obs.subscribe(event => {
-            let ts: number = Date.now();
-            if (ts - startTs > 5000) {
-                if (event.progress) {
-                    task.progress = event.progress;
-                }
-                task.status = event.status;
-                if (task.status == 1) task.progress = 1;
-                this.notificationService.pushLocalTask(task);
-            }
-        });
-    
-        return obs;
-    }
+    }  
 }
 
 export class DownloadSetup {
