+++ shanoir-ng-front/src/app/studies/shared/study.service.ts
@@ -13,10 +13,10 @@
  */
 import { HttpClient, HttpParams } from '@angular/common/http';
 import { Injectable, OnDestroy } from '@angular/core';
-import { Subscription } from 'rxjs';
-import { Observable } from 'rxjs';
+import { Observable, Subscription } from 'rxjs';
 
-import { Task, TaskState } from 'src/app/async-tasks/task.model';
+import { TaskState } from 'src/app/async-tasks/task.model';
+import { SingleDownloadService } from 'src/app/shared/mass-download/single-download.service';
 import { Tag } from 'src/app/tags/tag.model';
 import { DataUserAgreement } from '../../dua/shared/dua.model';
 import { EntityService } from '../../shared/components/entity/entity.abstract.service';
@@ -36,7 +36,6 @@ import {
     SubjectWithSubjectStudyDTO
 } from './study.dto';
 import { Study } from './study.model';
-import { MassDownloadService } from 'src/app/shared/mass-download/mass-download.service';
 
 @Injectable()
 export class StudyService extends EntityService<Study> implements OnDestroy {
@@ -47,8 +46,9 @@ export class StudyService extends EntityService<Study> implements OnDestroy {
     fileUploads: Map<number, Promise<void>> = new Map(); // current uploads
     private studyVolumesCache: Map<number, StudyStorageVolumeDTO> = new Map();
 
-    constructor(protected http: HttpClient, private keycloakService: KeycloakService, private studyDTOService: StudyDTOService) {
-        super(http)
+    constructor(protected http: HttpClient, private keycloakService: KeycloakService, private studyDTOService: StudyDTOService, 
+            private downloadService: SingleDownloadService) {
+        super(http);
     }
 
     getEntityInstance() { return new Study(); }
@@ -173,7 +173,7 @@ export class StudyService extends EntityService<Study> implements OnDestroy {
 
     downloadProtocolFile(fileName: string, studyId: number, state?: TaskState) {
         const endpoint = this.API_URL + '/protocol-file-download/' + studyId + "/" + fileName + "/";
-        //return this.downloadService.downloadSingleFile(endpoint, null, state);
+        return this.downloadService.downloadSingleFile(endpoint, null, state);
     }
 
     buildProtocolFileUrl(fileName: string, studyId: number): string {
@@ -182,7 +182,7 @@ export class StudyService extends EntityService<Study> implements OnDestroy {
 
     downloadDuaFile(fileName: string, studyId: number, state?: TaskState) {
         const endpoint = this.API_URL + '/dua-download/' + studyId + "/" + fileName + "/";
-        //return this.downloadService.downloadSingleFile(endpoint, null, state);
+        return this.downloadService.downloadSingleFile(endpoint, null, state);
     }
 
     downloadDuaBlob(fileName: string, studyId: number): Promise<Blob> {
